% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M2}
\alias{M2}
\title{69 trapezoidal fuzzy numbers.}
\format{
A matrix with 69 rows and 4 columns:
\describe{
  \item{inf0}{infimum of support of the trapezoidal fuzzy numbers, real number}
  \item{inf1}{infimum of core of the trapezoidal fuzzy numbers, real number}
  \item{sup1}{supremum of core of the trapezoidal fuzzy numbers, real number}
  \item{sup0}{supremum of support of the trapezoidal fuzzy numbers, real number}
}
}
\source{
\url{http://bellman.ciencias.uniovi.es/SMIRE/FuzzyRatingScaleQuestionnaire-SanIgnacio.html}
}
\usage{
M2
}
\description{
A dataset containing 69 trapezoidal fuzzy numbers.
The data correspond to the well-knows questionaire TIMSS-PIRLS2011.
This questionnaire was adapted to allow a double-type response, namely, the
original Likert and a fuzzy rating scale-based (to simplify, trapezoidal).
The questionnaire was conducted on 69 fourth grade students from Colegio San
Ignacio (Oviedo-Asturias, Spain).
Trapezoidal fuzzy rating responses to the Question M2 "My teacher is easy to understand"
are collected in this dataset.
}
\examples{
M2
}
\references{
[1] Gil, M.A.; Lubiano, M.A.; De la Rosa de Saa, S.; Sinova, B.: Analyzing data from a
fuzzy rating scale-based questionnaire. A case study, Psicothema 27(2), 182-191 (2015)

[2] Lubiano, M.A.; De la Rosa de Saa, S.; Montenegro, M.; Sinova, B.; Gil, M.A.:
Descriptive analysis of responses to items in questionnaires. Why not a fuzzy rating
scale?, Information Sciences 360, 131-148 (2016)

[3] Lubiano, M.A.; Montenegro, M.; Sinova, B.; De la Rosa de Saa, S.; Gil, M.A.: Hypothesis
testing for means in connection with fuzzy rating scale-based data: algorithms and
applications, European Journal of Operational Research 251, 918-929 (2016)

[3] Lubiano, M.A.; Salas, A.; Carleos, C.; De la Rosa de Sáa, S.; Gil, M.Á.: Hypothesis
testing-based comparative analysis between rating scales for intrinsically imprecise data,
International Journal of Approximate Reasoning 88, 128-147 (2017)

[4] Lubiano, M.A.; Salas, A.; Gil, M.Á.: A hypothesis testing-based discussion on the
sensitivity of means of Fuzzy data with respect to data shape, Fuzzy Sets and Systems
328(1), 54-69 (2017)
}
\keyword{datasets}
